day6

+++ sky-server/src/main/java/com/sky/service/CategoryService.java
@@ -2,11 +2,17 @@ package com.sky.service;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.entity.Dish;
+import com.sky.entity.Setmeal;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,6 +25,10 @@ public class CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private DishMapper dishMapper;
+    @Autowired
+    private SetmealMapper setmealMapper;
 
     public void createCategory(CategoryDTO categoryDTO) {
         Category category = new Category();
@@ -28,6 +38,14 @@ public class CategoryService {
     }
 
     public void deleteCategory(Long id) {
+        List<Dish> dishes = dishMapper.selectDishesByCategoryId(id);
+        if (dishes != null && !dishes.isEmpty()) {
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+        List<Setmeal> setmeals = setmealMapper.selectSetmealsByCategoryId(id);
+        if (setmeals != null && !setmeals.isEmpty()) {
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
         categoryMapper.deleteCategory(id);
     }
